Use the configured dtype for the scale-fit normalisation constant
The scale fit computed its normalisation constant 1/sum(F_obs^2) via
.double(), which raises on MPS: the backend has no float64, and
PYTORCH_ENABLE_MPS_FALLBACK covers unimplemented ops, not unsupported
dtypes. This broke every test reaching get_scales()/refine_scaler() on
the Accelerator CI job, and equally breaks a real TORCHREF_DEVICE=mps
refinement.

The float64 was unnecessary rather than misplaced. The constant rescales
every term of the objective identically, so its own precision cancels out
of the minimiser, the gradient direction and the likelihood:penalty
ratio; it only has to be finite, positive and O(1)-producing. float32
gives it to ~5e-8 relative on real work sets (torch.sum reduces pairwise,
so error grows like log(N)*eps, not N*eps).

Also stop the batched direct-summation paths allocating their accumulator
as complex128 regardless of configuration, which raises on MPS once
max_memory_gb splits the reflection loop and made the batched output
dtype disagree with the unbatched one.

+++ torchref/scaling/scaler_base.py
@@ -634,10 +634,14 @@ class ScalerBase(DeviceMixin, DebugMixin, nn.Module):
         # enough out that ``sum(U**2)`` overflows. Dividing by ``sum(F_obs**2)`` makes the
         # loss dimensionless and O(R**2) whatever the structure size or data scale.
         #
-        # The reduction is float64 for precision over ~1e5 terms; the constant is fixed
-        # for the whole fit.
+        # Computed in the configured dtype. Because the constant rescales every term
+        # identically, its own precision does not enter the result: it only has to be
+        # finite, positive and of the right magnitude. float32 gives it to ~5e-8 relative
+        # (``sum`` reduces pairwise, so error grows like log(N)*eps, not N*eps), which
+        # cancels out of the minimiser and the gradient direction alike. Do not cast to
+        # float64 -- MPS has none and raises on ``.double()``.
         _f_obs_work = self._data.work.F.detach()
-        _norm = float(1.0 / _f_obs_work.double().pow(2).sum().clamp(min=1e-30))
+        _norm = float(1.0 / _f_obs_work.pow(2).sum().clamp(min=1e-30))
 
         class _ScalerXrayTarget(nn.Module):
             """The registry row, closed over the detached ``fcalc``."""
